NC vergleichen, nc zu Double, delete mehoden

+++ BewerbungHochschule/src/main/java/domain/Bewerber.java
@@ -6,14 +6,14 @@ import utils.Studiengang;
 public class Bewerber extends Person{
 
 	public Bewerber(String name, String vorname, int alter, int pid, String telefonnummer, String email,
-			Studiengang studiengang, Haertefall haertefall, int nc) {
+			Studiengang studiengang, Haertefall haertefall, double nc) {
 		super(name, vorname, alter, pid, telefonnummer, email, studiengang);
 		this.haertefall = haertefall;
 		this.nc = nc;
 	}
 
 	private Haertefall haertefall;
-	private int nc;
+	private double nc;
 	
 	
 	public Haertefall getHaertefall() {
@@ -22,7 +22,7 @@ public class Bewerber extends Person{
 	public void setHaertefall(Haertefall haertefall) {
 		this.haertefall = haertefall;
 	}
-	public int getNc() {
+	public double getNc() {
 		return nc;
 	}
 	public void setNc(int nc) {
